requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.13.4
    # via bs4
blinker==1.9.0
    # via streamlit
bs4==0.0.2
    # via jupiter (pyproject.toml)
cachetools==5.5.2
    # via
    #   google-auth
    #   streamlit
certifi==2025.6.15
    # via
    #   httpx
    #   requests
chardet==3.0.4
    # via httpx
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   ipython
    #   tqdm
comm==0.2.2
    # via ipykernel
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
debugpy==1.8.14
    # via ipykernel
decorator==5.2.1
    # via ipython
executing==2.2.0
    # via stack-data
faiss-cpu==1.11.0
    # via jupiter (pyproject.toml)
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.58.4
    # via matplotlib
fsspec==2025.5.1
    # via
    #   huggingface-hub
    #   torch
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.25.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.173.0
    # via google-generativeai
google-auth==2.40.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.5
    # via jupiter (pyproject.toml)
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
googletrans==4.0.0rc1
    # via jupiter (pyproject.toml)
grpcio==1.73.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.0
    # via google-api-core
h11==0.9.0
    # via httpcore
h2==3.2.0
    # via httpcore
hpack==3.0.0
    # via h2
hstspreload==2025.1.1
    # via httpx
httpcore==0.9.1
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.13.3
    # via googletrans
huggingface-hub==0.33.0
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
hyperframe==5.2.0
    # via h2
idna==2.10
    # via
    #   httpx
    #   requests
ipykernel==6.29.5
    # via jupiter (pyproject.toml)
ipython==9.3.0
    # via ipykernel
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
    #   torch
joblib==1.5.1
    # via scikit-learn
jsonschema==4.24.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.8.1
    # via
    #   ipykernel
    #   jupyter-client
kiwisolver==1.4.8
    # via matplotlib
langdetect==1.0.9
    # via jupiter (pyproject.toml)
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.3
    # via
    #   jupiter (pyproject.toml)
    #   seaborn
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mpmath==1.3.0
    # via sympy
narwhals==1.44.0
    # via
    #   altair
    #   plotly
nest-asyncio==1.6.0
    # via ipykernel
networkx==3.5
    # via torch
numpy==2.3.1
    # via
    #   contourpy
    #   faiss-cpu
    #   matplotlib
    #   pandas
    #   pydeck
    #   scikit-learn
    #   scipy
    #   seaborn
    #   streamlit
    #   transformers
packaging==25.0
    # via
    #   altair
    #   faiss-cpu
    #   huggingface-hub
    #   ipykernel
    #   matplotlib
    #   plotly
    #   streamlit
    #   transformers
pandas==2.3.0
    # via
    #   jupiter (pyproject.toml)
    #   seaborn
    #   streamlit
parso==0.8.4
    # via jedi
pillow==11.2.1
    # via
    #   matplotlib
    #   sentence-transformers
    #   streamlit
platformdirs==4.3.8
    # via jupyter-core
plotly==6.1.2
    # via jupiter (pyproject.toml)
prompt-toolkit==3.0.51
    # via ipython
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   streamlit
psutil==7.0.0
    # via ipykernel
pure-eval==0.2.3
    # via stack-data
pyarrow==20.0.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via google-generativeai
pydantic-core==2.33.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
pyparsing==3.2.3
    # via
    #   httplib2
    #   matplotlib
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pywin32==310
    # via jupyter-core
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
pyzmq==27.0.0
    # via
    #   ipykernel
    #   jupyter-client
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via transformers
requests==2.32.4
    # via
    #   jupiter (pyproject.toml)
    #   google-api-core
    #   huggingface-hub
    #   streamlit
    #   transformers
rfc3986==1.5.0
    # via httpx
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
safetensors==0.5.3
    # via transformers
scikit-learn==1.7.0
    # via
    #   jupiter (pyproject.toml)
    #   sentence-transformers
scipy==1.16.0
    # via
    #   scikit-learn
    #   sentence-transformers
seaborn==0.13.2
    # via jupiter (pyproject.toml)
sentence-transformers==4.1.0
    # via jupiter (pyproject.toml)
setuptools==80.9.0
    # via torch
six==1.17.0
    # via
    #   langdetect
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   httpcore
    #   httpx
soupsieve==2.7
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
streamlit==1.46.0
    # via jupiter (pyproject.toml)
sympy==1.14.0
    # via torch
tenacity==9.1.2
    # via streamlit
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.2
    # via transformers
toml==0.10.2
    # via streamlit
torch==2.7.1
    # via sentence-transformers
tornado==6.5.1
    # via
    #   ipykernel
    #   jupyter-client
    #   streamlit
tqdm==4.67.1
    # via
    #   google-generativeai
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
transformers==4.52.4
    # via sentence-transformers
typing-extensions==4.14.0
    # via
    #   altair
    #   beautifulsoup4
    #   google-generativeai
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sentence-transformers
    #   streamlit
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
watchdog==6.0.0
    # via streamlit
wcwidth==0.2.13
    # via prompt-toolkit
